refactor: migrate system pure types to party-display-core
