+++ app/src-tauri/party-display-core/src/lib.rs
@@ -1,1 +1,2 @@
 pub mod dlna_proxy;
+pub mod system;
